requirements.txt
tensorflow-gpu==2.10.0
numpy
opencv-python
pandas
matplotlib
scikit-image
tf-slim
tf-models-official